requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
pypdf
